MetaGraphsNext.jl

MetaGraphsNext.DOTFormat

MetaGraphsNext.MGFormat

MetaGraphsNext.MetaGraph

MetaGraphsNext.defaultweight

MetaGraphsNext.weightfunction

MetaGraphsNext.weighttype

MetaGraphsNext.DOTFormat — Type

struct DOTFormat <: AbstractGraphFormat end

If all metadata types support pairs or are nothing, you can save MetaGraphs in DOTFormat.

julia> using MetaGraphsNext

julia> using LightGraphs

julia> simple = MetaGraph(Graph());

julia> simple[:a] = nothing; simple[:b] = nothing; simple[:a, :b] = nothing;

julia> mktemp() do file, io
            savegraph(file, simple, DOTFormat())
            print(read(file, String))
        end
graph T {
    a
    b
    a -- b
}

julia> complicated = MetaGraph(DiGraph(),
            VertexMeta = Dict{Symbol, Int},
            EdgeMeta = Dict{Symbol, Int},
            gprops = (tagged = true,)
        );

julia> complicated[:a] = Dict(:code_1 => 1, :code_2 => 2);

julia> complicated[:b] = Dict(:code => 2);

julia> complicated[:a, :b] = Dict(:code => 12);

julia> mktemp() do file, io
            savegraph(file, complicated, DOTFormat())
            print(read(file, String))
        end
digraph G {
    tagged = true
    a [code_1 = 1, code_2 = 2]
    b [code = 2]
    a -> b [code = 12]
}

MetaGraphsNext.MGFormat — Type

struct MGFormat <: AbstractGraphFormat end

You can save MetaGraphs in a MGFormat, currently based on JLD2.

julia> using MetaGraphsNext

julia> using LightGraphs: Edge, Graph,  loadgraph, savegraph

julia> complicated = MetaGraph(Graph());

julia> mktemp() do file, io
            savegraph(file, complicated)
            loadgraph(file, "something", MGFormat()) == complicated
        end
true

MetaGraphsNext.MetaGraph — Method

MetaGraph(g;
    Label = Symbol,
    VertexMeta = nothing,
    EdgeMeta = nothing,
    gprops = nothing,
    weightfunction = eprops -> 1.0,
    defaultweight = 1.0
)

Construct a new meta graph based on g, where Label is the type of the vertex labels, VertexMeta is the type of the metadata at a vertex, and EdgeMeta is the type of the metadata at an edge. You can also attach arbitrary graph level metadata as gprops.

julia> using LightGraphs

julia> using MetaGraphsNext

julia> colors = MetaGraph(Graph(), VertexMeta = Int, EdgeMeta = Symbol, gprops = "special")
Meta graph based on a {0, 0} undirected simple Int64 graph with vertices indexed by Symbol(s), Int64(s) vertex metadata, Symbol(s) edge metadata, "special" as graph metadata, and default weight 1.0

Use setindex! to add a new vertex with the given metadata.

julia> colors[:red] = 1;

julia> colors[:yellow] = 2;

julia> colors[:blue] = 3;

You can access and change the metadata using indexing: zero arguments for graph metadata, one label for vertex metadata, and two labels for edge metadata.

julia> colors[]
"special"

julia> colors[:blue] = 4;

julia> colors[:blue]
4

julia> colors[:red, :yellow] = :orange;

julia> colors[:red, :yellow]
:orange

You can delete vertices and edges with delete!.

julia> delete!(colors, :red, :yellow);

julia> delete!(colors, :blue);

You can use the weightfunction keyword to specify a function which will transform vertex metadata into a weight. This weight must always be the same type as the defaultweight.

julia> weighted = MetaGraph(Graph(), EdgeMeta = Float64, weightfunction = identity);

julia> weighted[:red] = nothing; weighted[:blue] = nothing; weighted[:yellow] = nothing;

julia> weighted[:red, :blue] = 1.0; weighted[:blue, :yellow] = 2.0;

julia> the_weights = LightGraphs.weights(weighted)
metaweights

julia> size(the_weights)
(3, 3)

julia> the_weights[1, 3]
1.0

julia> diameter(weighted)
3.0

MetaGraphsNext inherit many methods from LightGraphs. In general, inherited methods refer to vertices by codes, not labels.

julia> is_directed(colors)
false

julia> nv(zero(colors))
0

julia> ne(copy(colors))
0

julia> add_vertex!(colors, :white, 4)
true

julia> add_edge!(colors, 1, 3, :pink)
true

julia> rem_edge!(colors, 1, 3)
true

julia> rem_vertex!(colors, 3)
true

julia> rem_vertex!(colors, 3)
false

julia> eltype(colors) == Int
true

julia> edgetype(colors) == Edge{Int}
true

julia> vertices(colors)
Base.OneTo(2)

julia> has_edge(colors, 1, 2)
false

julia> has_vertex(colors, 1)
true

julia> LightGraphs.SimpleGraphs.fadj(colors, 1) == Int[]
true

julia> LightGraphs.SimpleGraphs.badj(colors, 1) == Int[]
true

julia> colors == colors
true

julia> issubset(colors, colors)
true

julia> SimpleGraph(colors)
{2, 0} undirected simple Int64 graph

You can seemlessly make MetaGraphsNext based on DiGraphs as well.

julia> rock_paper_scissors = MetaGraph(DiGraph(), Label = Symbol, EdgeMeta = Symbol);

julia> rock_paper_scissors[:rock] = nothing; rock_paper_scissors[:paper] = nothing; rock_paper_scissors[:scissors] = nothing;

julia> rock_paper_scissors[:rock, :scissors] = :rock_beats_scissors; rock_paper_scissors[:scissors, :paper] = :scissors_beats_paper; rock_paper_scissors[:paper, :rock] = :paper_beats_rock;

julia> is_directed(rock_paper_scissors)
true

julia> haskey(reverse(rock_paper_scissors), :scissors, :rock)
true

julia> SimpleDiGraph(rock_paper_scissors)
{3, 3} directed simple Int64 graph

julia> sub_graph, _ = induced_subgraph(rock_paper_scissors, [1, 3]);

julia> haskey(sub_graph, :rock, :scissors)
true

julia> delete!(rock_paper_scissors, :paper);

julia> rock_paper_scissors[:rock, :scissors]
:rock_beats_scissors

MetaGraphsNext.defaultweight — Method

defaultweight(g)

Return the default weight for metagraph g.

julia> using MetaGraphsNext

julia> using LightGraphs: Graph

julia> defaultweight(MetaGraph(Graph(), defaultweight = 2))
2

MetaGraphsNext.weightfunction — Method

weightfunction(g)

Return the weight function for metagraph g.

julia> using MetaGraphsNext

julia> using LightGraphs: Graph

julia> weightfunction(MetaGraph(Graph(), weightfunction = identity))(0)
0

MetaGraphsNext.weighttype — Method

weighttype(g)

Return the weight type for metagraph g.

julia> using MetaGraphsNext

julia> using LightGraphs: Graph

julia> weighttype(MetaGraph(Graph(), defaultweight = 1.0))
Float64

LightGraphs.SimpleGraphs.add_edge! — Method

add_edge!(g, u, v, val)

Add an edge (u, v) to MetaGraph g having value val.

LightGraphs.SimpleGraphs.add_vertex! — Method

add_vertex!(g, label, val)

Add a vertex to MetaGraph g with label label having value val.

return true if the vertex has been added, false otherwise.