API reference

Basics

MetaGraphsNext.MetaGraph — Type

MetaGraph{T,Label,Graph,VertexMeta,EdgeMeta,GraphMeta,WeightFunction,U} <: AbstractGraph{T}

A graph type with custom vertex labels containing vertex-, edge- and graph-level metadata.

Vertex labels have type Label, while vertex (resp. edge, resp. graph) metadata has type VertexMeta (resp. EdgeMeta, resp. GraphMeta). It is recommended not to set Label to Int to avoid confusion between vertex labels and vertex codes (which have type T <: Integer).

Fields

g::Graph: underlying, data-less graph with vertex indices of type T

vprops::Dict{Label,Tuple{T,VertexMeta}}: dictionary mapping vertex labels to vertex codes and metadata

eprops::Dict{Tuple{Label,Label},EdgeMeta}: dictionary mapping edge labels such as (label_u, label_v) to edge metadata

gprops::GraphMeta: graph metadata

weightfunction::WeightFunction: function defining edge weight from edge metadata

defaultweight::U: default weight for the edges

metaindex::Dict{T,Label}: dictionary mapping vertex codes to vertex labels

MetaGraphsNext.MetaGraph — Method

MetaGraph(g;
    Label = Symbol,
    VertexMeta = nothing,
    EdgeMeta = nothing,
    gprops = nothing,
    weightfunction = eprops -> 1.0,
    defaultweight = 1.0
)

Base.haskey — Method

haskey(g, :v1, :v2)

Determine whether a graph g contains an edge from :v1 to :v2. The order of :v1 and :v2 only matters if g is a digraph.

Base.haskey — Method

haskey(g, :label)

Determine whether a graph g contains the vertex :label.

MetaGraphsNext.code_for — Method

code_for(meta::MetaGraph, vertex_label)

Find the code associated with a vertex_label. This can be useful to pass to methods inherited from Graphs. Note, however, that vertex codes could be reassigned after vertex deletion.

MetaGraphsNext.label_for — Method

label_for(meta::MetaGraph, vertex_code)

Find the label associated with a vertex_code. This can be useful to interpret the results of methods inherited from Graphs. Note, however, that vertex codes could be reassigned after vertex deletion.

MetaGraphsNext.defaultweight — Method

defaultweight(g)

Return the default weight for metagraph g.

MetaGraphsNext.weightfunction — Method

weightfunction(g)

Return the weight function for metagraph g.

MetaGraphsNext.weighttype — Method

weighttype(g)

Return the weight type for metagraph g.

Graphs.jl interface

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(g, u, v, val)

Add an edge (u, v) to MetaGraph g having value val.

Return true if the edge has been added, false otherwise.

Graphs.SimpleGraphs.add_vertex! — Method

add_vertex!(g, label, val)

Add a vertex to MetaGraph g with label label having value val.

Return true if the vertex has been added, false otherwise.

Reading / writing

MetaGraphsNext.DOTFormat — Type

struct DOTFormat <: AbstractGraphFormat end

If all metadata types support pairs or are nothing, you can save MetaGraphs in DOTFormat.

MetaGraphsNext.MGFormat — Type

struct MGFormat <: AbstractGraphFormat end

You can save MetaGraphs in a MGFormat, currently based on JLD2.